Separate tests for standard token exchange v2
Closes #37105
